@@ -12,43 +12,42 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^10.1.5",
-    "@angular/cdk": "^10.2.4",
-    "@angular/common": "^10.1.5",
-    "@angular/compiler": "^10.1.5",
-    "@angular/core": "^10.1.5",
+    "@angular/animations": "^10.2.3",
+    "@angular/cdk": "^10.2.7",
+    "@angular/common": "^10.2.3",
+    "@angular/compiler": "^10.2.3",
+    "@angular/core": "^10.2.3",
     "@angular/flex-layout": "^10.0.0-beta.32",
-    "@angular/forms": "^10.1.5",
-    "@angular/material": "^10.2.4",
-    "@angular/platform-browser": "^10.1.5",
-    "@angular/platform-browser-dynamic": "^10.1.5",
-    "@angular/router": "^10.1.5",
-    "@fortawesome/angular-fontawesome": "^0.6.1",
+    "@angular/forms": "^10.2.3",
+    "@angular/material": "^10.2.7",
+    "@angular/platform-browser": "^10.2.3",
+    "@angular/platform-browser-dynamic": "^10.2.3",
+    "@angular/router": "^10.2.3",
+    "@fortawesome/angular-fontawesome": "^0.7.0",
     "@fortawesome/fontawesome-svg-core": "^1.2.32",
     "@fortawesome/free-brands-svg-icons": "^5.15.1",
     "@fortawesome/free-regular-svg-icons": "^5.15.1",
     "@fortawesome/free-solid-svg-icons": "^5.15.1",
-    "chart.js": "^2.9.3",
+    "@swimlane/ngx-charts": "^16.0.0",
     "chartjs-plugin-datalabels": "^0.7.0",
     "fast-deep-equal": "^3.1.3",
     "handlebars": "^4.7.6",
     "lodash-es": "^4.17.15",
-    "ng2-charts": "^2.4.2",
     "rxjs": "^6.6.3",
     "tslib": "^2.0.3",
     "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.1000.8",
-    "@angular/cli": "^10.1.6",
-    "@angular/compiler-cli": "^10.1.5",
-    "@angular/language-service": "^10.1.5",
-    "@types/jasmine": "^3.5.14",
+    "@angular-devkit/build-angular": "^0.1100.1",
+    "@angular/cli": "^10.2.0",
+    "@angular/compiler-cli": "^10.2.3",
+    "@angular/language-service": "^10.2.3",
+    "@types/jasmine": "^3.6.1",
     "@types/jasminewd2": "^2.0.8",
-    "@types/lodash": "^4.14.161",
+    "@types/lodash": "^4.14.165",
     "@types/lodash-es": "^4.17.3",
-    "@types/node": "^14.11.8",
-    "codelyzer": "^5.2.2",
+    "@types/node": "^14.14.7",
+    "codelyzer": "^6.0.1",
     "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~5.0.0",
     "karma": "~5.0.0",
